third_party/blink/web_tests/platform/generic/dom/text/remove-expected.txt

This tests the DOM 4 remove method on a Text node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testDiv.childNodes.length is 1
PASS testDiv.childNodes.length is 0
PASS testDiv.childNodes.length is 0
PASS successfullyParsed is true

TEST COMPLETE
